pyproject.toml
[project]
name = "timetable-fastapi"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = "~=3.12.9"
readme = "README.md"
dependencies = [
    "pandas~=2.2.3",
    "fastapi~=0.115.8",
    "uvicorn~=0.34.0",
    "pulp~=3.0.2",
    "python-dotenv~=1.1.0",
    "gurobipy>=12.0.3",
]

[dependency-groups]
dev = [
    "ipykernel~=6.29.5",
    "pytest-mock~=3.14.0",
    "pytest~=8.3.3",
    "pytest-cov~=6.0.0",
    "requests~=2.32.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests",]

[tool.hatch.build.targets.wheel]
packages = ["src"]
